Rename superlunary's `RemoteError` to `Rig.Error`
Despite the name it has nothing to do with RPC. superlunary stages code —
`Rig.dispatch` compiles a quoted `Expr` at runtime and runs it in a `Rig`,
a separate classloader — so there is no service, no peer and no wire
protocol; the only thing shared with obligatory's `Rpc` is the word
"remote", and obligatory owns that name anyway.

What the error reports is the *marshalling* boundary: `Serialization`,
`Deserialization`, `Unknown`, raised entirely within `object Stageable` as
values cross into the rig's `Transport` and back. Downstream signatures in
sedentary now read `raises Compiler.Error raises Rig.Error`, which names
the two things that can go wrong instead of one of them twice over.

Also rewrites the proposal's "Kept whole after verification" section.
Twenty-six of its thirty-odd names have since moved, because the
verification behind it looked for a top-level type of the *same name* and
so missed opaque types, companions inside an `internal` object, platform
source directories mistaken for components, and hosts named differently
from the error's own prefix. `Io.Error` was even listed there as a name
that keeps its compound form.
