+++ doc/errors/306.md
@@ -1,6 +1,12 @@
-# `SN-306` — `RemoteError`
+# `SN-306` — `Rig.Error`
 
-Raised when Superlunary fails to perform a remote operation.
+Raised when Superlunary cannot move a value across the boundary into or
+out of a `Rig` — the classloader in which staged code is compiled and run.
+
+This is a failure of *marshalling*, not of the staged code itself: the
+value could not be written into the rig's `Transport` (`Json` or `Pojo`),
+or could not be read back out of it. A failure to compile the staged tree
+is `Compiler.Error`, which `Rig.dispatch` declares alongside this one.
 
 ## Cause
 
@@ -41,4 +47,4 @@ The cause does not match either of the more specific variants.
 
 ## See also
 
-- Source: `lib/superlunary/src/core/superlunary.RemoteError.scala`
+- Source: `lib/superlunary/src/core/superlunary.Rig.scala`
