+++ lib/sedentary/src/core/sedentary.Stress.scala
@@ -123,7 +123,7 @@ extends Rig:
             threading: Threading,
             suite:     Testable,
             codepoint: Codepoint )
-  :   Unit raises Compiler.Error raises RemoteError =
+  :   Unit raises Compiler.Error raises Rig.Error =
 
     val concurrency0: Optional[Int] = concurrency
     val start: Int = concurrency0.or(1)
