+++ lib/superlunary/src/core/soundness_superlunary_core.scala
@@ -32,4 +32,4 @@
                                                                                                   */
 package soundness
 
-export superlunary.{embeddings, Executor2, References, RemoteError, Rig, Stage, Stageable}
+export superlunary.{embeddings, Executor2, References, Rig, Stage, Stageable}
